% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lq.R
\name{Lq}
\alias{Lq}
\title{Q function lower bound in variational approximation that needs to be maximized in EM}
\usage{
Lq(s2, mu, alpha, v2, pi, Lambda, z)
}
\description{
Q function lower bound in variational approximation that needs to be maximized in EM
}
