% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XING_single.R
\name{XING_single}
\alias{XING_single}
\title{Starting model for X-ING}
\usage{
XING_single(
  z,
  Lambda,
  iterT = 20,
  eps_thres = 0.001,
  vk_init = 0.1,
  bound = 1e-04,
  pi_init = 1e-04
)
}
\arguments{
\item{z}{a matrix of association summary statistics with multivariate contexts.}

\item{Lambda}{a matrix of tissue-tissue correlation matrix among all tissues due to potential sample overlap.}

\item{iterT}{maximum iteration times.}

\item{eps_thres:}{stopping rule threshold.}

\item{vk_init:}{initial value for the variance of latent genetic association distribution.}

\item{pi_init:}{initial probability for having latent genetic associations.}
}
\description{
\code{X-ING} starting model without considering shared patterns across contexts and data types.
}
\details{
This function \code{XING_single}, X-ING estimates the posterior means of input statistics, models the latent binary association probabilities based on the posterior means, and outputs the posterior mean and probability of association for each input statistic without considering shared patterns across contexts and data types.
}
