% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XING.R
\name{XING}
\alias{XING}
\title{Cross INtegrative Genomics method (X-ING)}
\usage{
XING(
  z_list,
  Lambda_list = NULL,
  CC = 2,
  PC_list = NULL,
  iterT = 20,
  method = "CCAPCA",
  tolerance = 1e-08
)
}
\arguments{
\item{z_list}{a list of association summary statistics, each with multivariate contexts.}

\item{Lambda_list}{a list of tissue-tissue correlation matrix among all tissues due to potential sample overlap.}

\item{CC}{number of canonical components to keep in canonical correlation analysis (CCA).}

\item{PC_list}{numbers of principal components to keep in principal correlation analysis (PCA).}

\item{iterT}{maximum iteration times.}

\item{method}{methods to extract shared patterns.}

\item{tolerance}{tolerance for iteration of generalized canonical correlation analysis.}
}
\value{
a list with following elements:
\item{mu}{posterior means of input statistics.}
\item{alpha}{posterior probability of association for each input statistic.}
}
\description{
X-ING models and estimates the latent binary association status of each statistic, and captures the omic-shared, context-shared, omic-specific and context-specific major patterns in a hierarchical Bayesian model.
}
\details{
This function \code{XING}, jointly estimates the posterior means of input statistics, models the latent binary association probabilities based on the posterior means, and outputs the posterior mean and probability of association for each input statistic.
}
\examples{
library(MASS)
library(CCA)
library(RGCCA)
data(example)
z_list <- lapply(example,function(x) x[[1]])
res <- XING(z_list = z_list[1:2])
res <- XING(z_list = z_list)
}
